minor changes to fix some compile errors in port of latest hydro to indigo dumbed down error checking to deal with non-latest install of ceres

+++ industrial_extrinsic_cal/src/ros_camera_observer.cpp
@@ -185,7 +185,7 @@ int ROSCameraObserver::getObservations(CameraObservations &cam_obs)
       // updates to OpenCV
       std::vector<cv::KeyPoint> keypoints;
       circle_detector_ptr_->detect(image_roi_, keypoints);
-      ROS_DEBUG("found %d keypoints", keypoints.size());
+      ROS_DEBUG("found %d keypoints", (int)keypoints.size());
       if(successful_find){ // determine orientation, and sort points to correct correspondence
 	int lw_lt_index = pattern_rows_*pattern_cols_ - pattern_cols_; // lower right point's index
 	int lw_rt_index = pattern_rows_*pattern_cols_ -1; // lower left point's index
